vit-image-classification/requirements.txt
transformers
pillow

# https://pytorch.org/
torch
torchvision
torchaudio